services/ai-service/src/services/face_service.py
import os
import logging
import uuid
import base64
import random
from datetime import datetime
from typing import Dict, List, Optional
import numpy as np
from PIL import Image
import io

from ..models.face import FaceMatchResponse
from ..config import settings

logger = logging.getLogger(__name__)

class FaceService:
    def __init__(self):
        self.model_path = settings.FACE_MODEL_PATH
        logger.info(f"Initializing FaceService with model path: {self.model_path}")
        
        # In a real implementation, we would load the face recognition model here
        # For this demo, we'll simulate the AI analysis
    
    async def match_faces(self, document_id: str, selfie_id: str, verification_id: str, document_path: str, selfie_path: str) -> FaceMatchResponse:
        """
        Match a selfie with a document photo
        """
        logger.info(f"Matching faces: document_id={document_id}, selfie_id={selfie_id}")
        
        try:
            # In a real implementation, we would:
            # 1. Load both images
            # 2. Detect faces in both images
            # 3. Extract face embeddings
            # 4. Calculate similarity between embeddings
            # 5. Determine if they match
            
            # For this demo, we'll simulate the face matching
            is_match, confidence = self._match_faces_simulation()
            
            # Create response
            return FaceMatchResponse(
                id=str(uuid.uuid4()),
                verification_id=verification_id,
                document_id=document_id,
                selfie_id=selfie_id,
                is_match=is_match,
                confidence=confidence,
                created_at=datetime.now()
            )
        except Exception as e:
            logger.error(f"Error matching faces: {str(e)}")
            raise
    
    async def match_faces_base64(self, document_id: str, selfie_id: str, verification_id: str, document_image: str, selfie_image: str) -> FaceMatchResponse:
        """
        Match faces from base64 encoded images
        """
        logger.info(f"Matching faces from base64: document_id={document_id}, selfie_id={selfie_id}")
        
        try:
            # Decode base64 images
            document_data = base64.b64decode(document_image)
            selfie_data = base64.b64decode(selfie_image)
            
            # Save images to temporary files
            document_path = f"uploads/temp_doc_{document_id}.jpg"
            selfie_path = f"uploads/temp_selfie_{selfie_id}.jpg"
            
            with open(document_path, "wb") as f:
                f.write(document_data)
            
            with open(selfie_path, "wb") as f:
                f.write(selfie_data)
            
            # Match faces
            result = await self.match_faces(document_id, selfie_id, verification_id, document_path, selfie_path)
            
            # Clean up temporary files
            os.remove(document_path)
            os.remove(selfie_path)
            
            return result
        except Exception as e:
            logger.error(f"Error matching faces from base64: {str(e)}")
            raise
    
    def _match_faces_simulation(self) -> tuple:
        """
        Simulate face matching
        """
        # In a real implementation, we would use a face recognition model
        # For this demo, we'll return a random result with 85% chance of matching
        is_match = random.random() > 0.15
        
        if is_match:
            confidence = 75.0 + random.random() * 20.0  # 75-95% confidence for matches
        else:
            confidence = 30.0 + random.random() * 40.0  # 30-70% confidence for non-matches
        
        return is_match, confidence
