services/ai-service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
pillow>=9.5.0
numpy>=1.22.0,<1.24.0
pydantic>=1.10.7
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
httpx>=0.24.0
aiofiles>=23.1.0
python-dotenv>=1.0.0
